review comment
review comments


clean comments


CNAM-238 delete the added drug families definitions

it's better to do them later when the confluence page is finished
CNAM 238 code review


CNAM 238 intialize exception to empty sets


CNAM 238 add caching


CNAM 238 remove comments


CNAM 238 replace lists bu sets

clean imports
CNAM 238 refactor and clean review


CNAM 238 refactor naiveDrugs

review refactoring
CNAM 238 comment atc drug definition


CNAM-238 antidepresseurs drug extractor


CNAM-238 refactoring


CNAM-238 drug extractor refacoring


CNAM 238 clean code


CNAM 238 clean code


CNAM 238 cleaning code


CNAM 238 refactor naiveDrugs


CNAM-238 first version
