+++ src/main/scala/fr/polytechnique/cmap/cnam/etl/extractors/drugs/DrugConfig.scala
@@ -1,7 +1,8 @@
 package fr.polytechnique.cmap.cnam.etl.extractors.drugs
 
+
 trait DrugConfig {
 
   val name: String
-  val cip13Codes: List[String]
+  val cip13Codes: Set[String]
 }
