+++ src/main/scala/fr/polytechnique/cmap/cnam/etl/extractors/drugs/DrugPurchases.scala
@@ -3,10 +3,10 @@ package fr.polytechnique.cmap.cnam.etl.extractors.drugs
 import org.apache.spark.sql.Dataset
 import fr.polytechnique.cmap.cnam.etl.events.{Drug, Event}
 
+
 trait DrugPurchases {
 
-  val drugName: String
-  val drugCodes: List[String]
+  val drugCodes: Map[String, Set[String]]
 
   def extract: Dataset[Event[Drug]]
 }
